moving docs to src
