+++ src/cmd/singularity/cli/build.go
@@ -15,7 +15,7 @@ import (
 	"github.com/singularityware/singularity/src/pkg/sylog"
 	"github.com/spf13/cobra"
 
-	"github.com/singularityware/singularity/docs"
+	"github.com/singularityware/singularity/src/docs"
 )
 
 var (
