+++ src/cmd/singularity/cli/capability.go
@@ -8,7 +8,7 @@ package cli
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/singularityware/singularity/docs"
+	"github.com/singularityware/singularity/src/docs"
 )
 
 func init() {
