+++ src/cmd/singularity/cli/instance_start.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/singularityware/singularity/docs"
+	"github.com/singularityware/singularity/src/docs"
 )
 
 func init() {
