+++ src/cmd/singularity/cli/pull.go
@@ -13,7 +13,7 @@ import (
 	"github.com/singularityware/singularity/src/pkg/sylog"
 	"github.com/spf13/cobra"
 
-	"github.com/singularityware/singularity/docs"
+	"github.com/singularityware/singularity/src/docs"
 )
 
 var (
