+++ src/cmd/singularity/cli/singularity.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/singularityware/singularity/docs"
+	"github.com/singularityware/singularity/src/docs"
 )
 
 // Global variables for singularity CLI
