+++ src/docs/2.x-tests/01-sanity.sh
