+++ src/docs/2.x-tests/23-bootstrap_apps.sh
